+++ src/main/java/com/thigas/quack/domain/repository/IUserRepository.java
@@ -7,9 +7,9 @@ import com.thigas.quack.domain.entity.UserEntity;
 public interface IUserRepository {
     UserEntity save(UserEntity user);
 
-    Optional<UserEntity> findById(Long id);
+    Optional<UserEntity> findById(int id);
 
     Iterable<UserEntity> findAll();
 
-    void deleteById(Long id);
+    void deleteById(int id);
 }
