+++ src/main/java/com/thigas/quack/infrastructure/persistence/entity/UserAchievementsModel.java
@@ -15,7 +15,7 @@ public class UserAchievementsModel {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     @ManyToOne
     @JoinColumn(name = "user_id", nullable = false)
@@ -26,16 +26,16 @@ public class UserAchievementsModel {
     private AchievementModel achievement;
 
     /**
-     * @return Long return the id
+     * @return int return the id
      */
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
     /**
      * @param id the id to set
      */
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
